requirements.txt
flask
pillow
pandas
openpyxl
werkzeug
reportlab
gunicorn